requirements/project.txt
Django==1.4
Mezzanine==1.0.10
Pillow==1.7.7
South==0.7.4
bleach==1.1.1
distribute==0.6.24
django-form-utils==0.2.0
filebrowser-safe==0.2.6
grappelli-safe==0.2.6
html5lib==0.95
simplejson==2.5.0
wsgiref==0.1.2
